+++ src/games/asteroid/src/interfaces/asteroid.interface.ts
@@ -1,0 +1,5 @@
+export interface IAsteroidConstructor {
+  scene: Phaser.Scene;
+  size: number;
+  options?: Phaser.Types.GameObjects.Graphics.Options;
+}