+++ src/games/asteroid/src/interfaces/graphics.interface.ts
@@ -1,0 +1,4 @@
+export interface IGraphicsConstructor {
+  scene: Phaser.Scene;
+  options?: Phaser.Types.GameObjects.Graphics.Options;
+}